Fix coachreport import error Fixes #3224, which was caused by an unprefaced import statement & subsequent name error. Additionally, several inline imports were, I think, added as a sneaky way to circumvent a circular import... that is coachreports.views was importing control_panel.api_resources, and vice-versa. So I just moved the offending function out of coachreports.views into the facility package, since that was its only dependency.
