+++ kalite/coachreports/views.py
@@ -16,6 +16,8 @@ from django.utils.translation import ungettext, ugettext_lazy, ugettext as _
 from .api_views import get_data_form, stats_dict
 from django.shortcuts import get_object_or_404
 
+from kalite.facility import get_accessible_objects_from_logged_in_user
+from kalite.control_panel import api_resources as control_panel_api_resources
 from kalite.distributed.api_views import compute_total_points
 from kalite.facility.decorators import facility_required
 from kalite.facility.models import Facility, FacilityUser, FacilityGroup
@@ -30,55 +32,6 @@ from kalite.topic_tools import get_topic_exercises, get_topic_videos, get_knowle
 
 # shared by test_view and test_detail view
 SUMMARY_STATS = [ugettext_lazy('Max'), ugettext_lazy('Min'), ugettext_lazy('Average'), ugettext_lazy('Std Dev')]
-
-
-def get_accessible_objects_from_logged_in_user(request, facility):
-    """Given a request, get all the facility/group/user objects relevant to the request,
-    subject to the permissions of the user type.
-
-    Make sure the returned `facilities` object is always a Facility queryset or an empty list.
-    """
-
-    # Options to select.  Note that this depends on the user.
-    facilities = []
-    if request.user.is_superuser:
-        facilities = Facility.objects.all()
-        # Groups is now a list of objects with a key for facility id, and a key
-        # for the list of groups at that facility.
-        # TODO: Make this more efficient.
-        groups = [{"facility": f.id, "groups": FacilityGroup.objects.filter(facility=f)} for f in facilities]
-
-    elif "facility_user" in request.session:
-        user = request.session["facility_user"]
-        if user.is_teacher:
-            facilities = Facility.objects.all()
-            groups = [{"facility": f.id, "groups": FacilityGroup.objects.filter(facility=f)} for f in facilities]
-        else:
-            # Students can only access their group
-            if facility and isinstance(facility, Facility):
-                facilities = Facility.objects.filter(id=facility.id)
-            if not user.group:
-                groups = []
-            else:
-                groups = [{"facility": user.facility.id,
-                           "groups": FacilityGroup.objects.filter(id=request.session["facility_user"].group)}]
-    elif facility:
-        facilities = Facility.objects.filter(id=facility.id)
-        groups = [{"facility": facility.id, "groups": FacilityGroup.objects.filter(facility=facility)}]
-    else:
-        # defaults to all facilities and groups
-        facilities = Facility.objects.all()
-        groups = [{"facility": f.id, "groups": FacilityGroup.objects.filter(facility=f)} for f in facilities]
-
-    ungrouped_available = False
-    for f in facilities:
-        # Check if there is at least one facility with ungrouped students.
-        ungrouped_available = f.has_ungrouped_students
-        if ungrouped_available:
-            break
-
-    return (groups, facilities, ungrouped_available)
-
 
 def plotting_metadata_context(request, facility=None, topic_path=[], *args, **kwargs):
     """Basic context for any plot: get the data form, a dictionary of stat definitions,
@@ -234,8 +187,7 @@ def tabular_view(request, report_type="exercise"):
 
     if group_id:
         # Narrow by group
-        from control_panel.api_resources import UNGROUPED_KEY
-        if group_id == UNGROUPED_KEY:
+        if group_id == control_panel_api_resources.UNGROUPED_KEY:
             users = FacilityUser.objects.filter(group__isnull=True, is_teacher=False)
             if facility:
                 # filter only those ungrouped students for the facility
@@ -378,11 +330,10 @@ def test_view(request):
     # Get the TestLog objects generated by this group of students
     # TODO(cpauya): what about queryset for ungrouped students?
     test_logs = None
-    from control_panel.api_resources import UNGROUPED_KEY
     if group_id:
         test_logs = TestLog.objects.filter(user__group=group_id)
         # Narrow all by ungroup facility user
-        if group_id == UNGROUPED_KEY:
+        if group_id == control_panel_api_resources.UNGROUPED_KEY:
             test_logs = TestLog.objects.filter(user__group__isnull=True)
             if facility:
                 TestLog.objects.filter(user__facility=facility, user__group__isnull=True)
@@ -501,11 +452,10 @@ def test_detail_view(request, test_id):
 
     # get all of the test logs for this specific test object and generated by these specific users
     if group_id:
-        from control_panel.api_resources import UNGROUPED_KEY
         test_logs = TestLog.objects.filter(user__group=group_id, test=test_id)
 
         # Narrow all by ungroup facility user
-        if group_id == UNGROUPED_KEY:
+        if group_id == control_panel_api_resources.UNGROUPED_KEY:
             if facility:
                 test_logs = TestLog.objects.filter(user__group__isnull=True)
             else:
@@ -644,13 +594,12 @@ def get_user_queryset(request, facility, group_id):
     """Return set of users appropriate to the facility and group"""
     student_ordering = ["last_name", "first_name", "username"]
     (groups, facilities, ungrouped_available) = get_accessible_objects_from_logged_in_user(request, facility=facility)
-    from control_panel.api_resources import UNGROUPED_KEY
     if group_id:
         # Narrow by group
         users = FacilityUser.objects.filter(
             group=group_id, is_teacher=False).order_by(*student_ordering)
         # Narrow all by ungroup user
-        if group_id == UNGROUPED_KEY:
+        if group_id == control_panel_api_resources.UNGROUPED_KEY:
             users = FacilityUser.objects.filter(group__isnull=True, is_teacher=False).order_by(*student_ordering)
             if facility:
                 users = FacilityUser.objects.filter(facility=facility, group__isnull=True,
