+++ kalite/control_panel/api_resources.py
@@ -13,7 +13,7 @@ from tastypie.resources import Resource, ModelResource
 
 from securesync.models import Zone, Device, SyncSession
 
-from kalite.coachreports.views import get_accessible_objects_from_logged_in_user
+from kalite.facility import get_accessible_objects_from_logged_in_user
 from kalite.facility.models import Facility, FacilityGroup, FacilityUser
 from kalite.main.models import AttemptLog, ExerciseLog
 from kalite.shared.api_auth import ObjectAdminAuthorization
